style: add bootstrap classes to buttons and tables in actor and movie CRUD pages feat: add confirmation dialog before deleting an entity feat: add formatted release date to movie list display feat: add "Add Actor" and "Add Movie" buttons to actors and movies pages, respectively feat: add labels to actor form inputs

+++ static/js/entity_crud.js
@@ -109,10 +109,12 @@ class EntityCRUD {
     }
 
     async delete(entityId) {
-        await fetch(`/api/${this.entityName}/${entityId}`, {
-            method: 'DELETE',
-        });
-        this.fetchAndDisplay();
+        if (confirm("Are you sure you want to delete this item?")) {
+            await fetch(`/api/${this.entityName}/${entityId}`, {
+                method: 'DELETE',
+            });
+            this.fetchAndDisplay();
+        }
     }
 
     async nextPage() {
